add a cache layer
The cache layer is going to automatically serialise/deserialise(stringyfy, parse) objects so we don’t have to call JSON.stringyfy/JSON.parse all the time. Our cache is also going to apply a timestamp to each item, and then figure out if the item is expired or not. If expired, it can automatically clean up async storage and free up som space.